requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.8
django-oauth-toolkit==1.3.2
django-oauth2-provider==0.2.6.1
djangorestframework==3.11.0
djangorestframework-oauth==1.1.0
idna==2.10
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.4.0
pytz==2020.1
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
